SWDEV-492272 [AMDSMI] Build/Compiler warnings messages
Fix compiler warnings



[ROCm/amdsmi commit: d0a7332d3257928886b494c13cd92089e34ee847]

+++ rocm_smi/include/rocm_smi/rocm_smi_utils.h
@@ -427,7 +427,7 @@ class TagTextContents_t
 
         decltype(auto) get_structured_data_subkey_last(const PrimaryKeyType& prim_key) {
             return (get_structured_value_by_keys(prim_key, get_structured_data_subkey_by_position(prim_key,
-                                                                                                  (get_structured_subkeys_size(prim_key) - 1))));
+                                                                                                  static_cast<int>((get_structured_subkeys_size(prim_key) - 1)))));
         }
 
         void reset() {
